Rename `CreateFacilityOrganizationSheet.tsx` to `FacilityOrganizationFormSheet.tsx` (#11901)
